refactor: move extra components out of answer option file
