+++ src/editors/sharedComponents/Feedback/index.jsx
@@ -1,0 +1,2 @@
+export { default as FeedbackBox } from './FeedbackBox';
+export { default as FeedbackControl } from './FeedbackControl';
